requirements.txt
azure-core==1.15.0
azure-storage-blob==12.8.1
certifi==2021.5.30        
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
cycler==0.10.0
et-xmlfile==1.1.0
feature-engine==0.6.0
idna==2.10
isodate==0.6.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
msrest==0.6.21
numpy==1.20.3
oauthlib==3.1.1
openpyxl==3.0.7
pandas==1.2.4
patsy==0.5.1
Pillow==8.2.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
statsmodels==0.12.2
threadpoolctl==2.1.0
urllib3==1.26.5